fix: removes circular dependencies in pn-pa-webapp and pn-personafisica
