+++ packages/pn-pa-webapp/src/App.tsx
@@ -20,18 +20,20 @@ import { MIXPANEL_TOKEN } from "./utils/constants";
 import Router from './navigation/routes';
 import { AUTH_ACTIONS, getOrganizationParty, logout } from './redux/auth/actions';
 import { useAppDispatch, useAppSelector } from './redux/hooks';
-import { RootState } from './redux/store';
+import { RootState, store } from './redux/store';
 import { getMenuItems } from './utils/role.utility';
 
 import { PAGOPA_HELP_EMAIL, SELFCARE_BASE_URL, VERSION } from './utils/constants';
 import { trackEventByType } from './utils/mixpanel';
 import { TrackEventType } from './utils/events';
 import './utils/onetrust';
+import { setUpInterceptor } from "./api/interceptors";
 
 const App = () => {
   useUnload(() => {
     trackEventByType(TrackEventType.APP_UNLOAD);
   });
+  setUpInterceptor(store);
 
   const loggedUser = useAppSelector((state: RootState) => state.userState.user);
   const loggedUserOrganizationParty = useAppSelector(
