+++ packages/pn-pa-webapp/src/api/auth/__test__/Auth.api.test.ts
@@ -2,7 +2,7 @@ import MockAdapter from 'axios-mock-adapter';
 
 import { AuthApi } from '../Auth.api';
 import { userResponse } from '../../../redux/auth/__test__/test-utils';
-import { authClient } from '../../axios';
+import { authClient } from '../../clients';
 import { AUTH_TOKEN_EXCHANGE } from '../auth.routes';
 
 export async function mockedExchangeToken() {
