--- packages/pn-pa-webapp/src/api/axios.ts
@@ -1,26 +1,0 @@
-import axios from 'axios';
-import { store } from '../redux/store';
-
-import { API_BASE_URL } from '../utils/constants';
-
-export const authClient = axios.create({
-  baseURL: API_BASE_URL,
-});
-
-export const apiClient = axios.create({
-  baseURL: API_BASE_URL,
-});
-
-apiClient.interceptors.request.use(
-  (config) => {
-    /* eslint-disable functional/immutable-data */
-    const token = store.getState().userState.user.sessionToken;
-    if (token && config.headers) {
-      config.headers.Authorization = 'Bearer ' + token;
-    }
-    return config;
-  },
-  (error) => Promise.reject(error)
-);
-
-export const externalClient = axios.create();
