+++ packages/pn-pa-webapp/src/api/external-registries/__test__/External-registries.api.test.ts
@@ -2,7 +2,7 @@ import MockAdapter from 'axios-mock-adapter';
 
 import { Party } from '../../../models/party';
 import { mockAuthentication } from '../../../redux/auth/__test__/test-utils';
-import { apiClient } from '../../axios';
+import { apiClient } from '../../clients';
 import { GET_PARTY_FOR_ORGANIZATION } from '../external-registries-routes';
 import { ExternalRegistriesAPI } from '../External-registries.api';
