+++ packages/pn-personafisica-webapp/src/App.tsx
@@ -25,13 +25,14 @@ import Router from './navigation/routes';
 import { logout } from './redux/auth/actions';
 import { useAppDispatch, useAppSelector } from './redux/hooks';
 import { MIXPANEL_TOKEN, PAGOPA_HELP_EMAIL, VERSION } from './utils/constants';
-import { RootState } from './redux/store';
+import { RootState, store } from './redux/store';
 import { Delegation } from './redux/delegation/types';
 import { getDomicileInfo, getSidemenuInformation } from './redux/sidemenu/actions';
 import { trackEventByType } from './utils/mixpanel';
 import { TrackEventType } from './utils/events';
 import './utils/onetrust';
 import { goToLoginPortal } from "./navigation/navigation.utility";
+import { setUpInterceptor } from "./api/interceptors";
 
 // TODO: get products list from be (?)
 const productsList: Array<ProductSwitchItem> = [
@@ -44,6 +45,7 @@ const productsList: Array<ProductSwitchItem> = [
 ];
 
 const App = () => {
+  setUpInterceptor(store);
   const dispatch = useAppDispatch();
   const { t, i18n } = useTranslation(['common', 'notifiche']);
   const loggedUser = useAppSelector((state: RootState) => state.userState.user);
