+++ packages/pn-personafisica-webapp/src/api/auth/Auth.api.ts
@@ -1,5 +1,5 @@
 import { User } from "../../redux/auth/types";
-import { authClient } from "../axios";
+import { authClient } from "../clients";
 import { AUTH_TOKEN_EXCHANGE } from "./auth.routes";
 
 export const AuthApi = {
