+++ packages/pn-personafisica-webapp/src/api/consents/Consents.api.ts
@@ -1,5 +1,5 @@
 import { Consent, ConsentActionType, ConsentType } from '../../models/consents';
-import { apiClient } from '../axios';
+import { apiClient } from '../clients';
 import { CONSENTS } from './consents.routes';
 
 export const ConsentsApi = {
