+++ packages/pn-personafisica-webapp/src/api/notifications/Notifications.api.ts
@@ -12,7 +12,7 @@ import { AxiosResponse } from 'axios';
 import { Delegator } from '../../redux/delegation/types';
 import { parseNotificationDetailForRecipient } from '../../utils/notification.utility';
 import { NotificationDetailForRecipient } from '../../models/NotificationDetail';
-import { apiClient } from '../axios';
+import { apiClient } from '../clients';
 import {
   NOTIFICATIONS_LIST,
   NOTIFICATION_DETAIL,
