+++ packages/pn-personafisica-webapp/src/api/notifications/__test__/Notifications.api.test.ts
@@ -8,7 +8,7 @@ import {
   formatToTimezoneString,
   getNextDay,
 } from '@pagopa-pn/pn-commons';
-import { apiClient } from '../../axios';
+import { apiClient } from '../../clients';
 import { NotificationsApi } from '../Notifications.api';
 import {
   notificationsFromBe,
